Adding sttp MonadError instance for cats MonadError
